General clean up continued, Docker improved, tests running in docker container
