+++ docker-compose.yml
@@ -1,0 +1,8 @@
+version: "3.9"
+services:
+  mixmasta:
+    container_name: mixmasta-container
+    command: tail -f /dev/null #Keeps container up on start, no other function
+    build: ./
+    volumes:
+      - $PWD:/mixmasta
