+++ mixmasta/constants.py
@@ -1,0 +1,17 @@
+# Constants used in multiple files
+COL_ORDER = [
+    "timestamp",
+    "country",
+    "admin1",
+    "admin2",
+    "admin3",
+    "lat",
+    "lng",
+    "feature",
+    "value",
+]
+
+GEO_TYPE_COUNTRY = "country"
+GEO_TYPE_ADMIN1 = "state/territory"
+GEO_TYPE_ADMIN2 = "county/district"
+GEO_TYPE_ADMIN3 = "municipality/town"
