+++ tests/test_mixmasta.py
@@ -5,23 +5,23 @@
 
 import unittest
 import warnings
-
-from click.testing import CliRunner
 import subprocess
 import logging
 import json
-from mixmasta import cli, mixmasta
-from pandas.util.testing import assert_frame_equal, assert_dict_equal
-import pandas as pd
 import gc
 import os
 
-if os.name == 'nt':
-    sep = '\\'
+from mixmasta import mixmasta
+from pandas.util.testing import assert_frame_equal, assert_dict_equal
+import pandas as pd
+
+if os.name == "nt":
+    sep = "\\"
 else:
-    sep = '/'
+    sep = "/"
 
 logger = logging.getLogger(__name__)
+
 
 class TestMixmaster(unittest.TestCase):
     """Tests for `mixmasta` package."""
@@ -30,7 +30,7 @@ class TestMixmaster(unittest.TestCase):
         """Set up test fixtures, if any."""
 
         # turn off warnings unless refactoring.
-        warnings.simplefilter('ignore')
+        warnings.simplefilter("ignore")
 
     def tearDown(self):
         """Tear down test fixtures, if any."""
@@ -42,7 +42,7 @@ class TestMixmaster(unittest.TestCase):
             pass
 
         try:
-           os.remove(f"outputs{sep}unittests_str.parquet.gzip")
+            os.remove(f"outputs{sep}unittests_str.parquet.gzip")
         except:
             pass
 
@@ -52,24 +52,34 @@ class TestMixmaster(unittest.TestCase):
         # Define mixmasta inputs:
         mp = f"inputs{sep}test1_input.json"
         fp = f"inputs{sep}test1_input.csv"
-        geo = 'admin2'
+        geo = "admin2"
         outf = f"outputs{sep}unittests"
 
         # Process:
         df, dct = mixmasta.process(fp, mp, geo, outf)
 
         # Load expected output:
-        output_df = pd.read_csv(f'outputs{sep}test1_output.csv', index_col=False)
+        output_df = pd.read_csv(f"outputs{sep}test1_output.csv", index_col=False)
         output_df = mixmasta.optimize_df_types(output_df)
-        with open(f'outputs{sep}test1_dict.json') as f:
+        with open(f"outputs{sep}test1_dict.json") as f:
             output_dict = json.loads(f.read())
 
         # Sort both data frames and reindex for comparison,.
-        cols = ['timestamp','country','admin1','admin2','admin3','lat','lng','feature','value']
+        cols = [
+            "timestamp",
+            "country",
+            "admin1",
+            "admin2",
+            "admin3",
+            "lat",
+            "lng",
+            "feature",
+            "value",
+        ]
         df.sort_values(by=cols, inplace=True)
         output_df.sort_values(by=cols, inplace=True)
         df.reset_index(drop=True, inplace=True)
-        output_df.reset_index(drop =True, inplace=True)
+        output_df.reset_index(drop=True, inplace=True)
 
         # Assertions
         assert_frame_equal(df, output_df)
@@ -83,122 +93,159 @@ class TestMixmaster(unittest.TestCase):
         """
 
         # Define mixmasta inputs:
-        mp = f'inputs{sep}test2_assetwealth_input.json'
-        fp = f'inputs{sep}test2_assetwealth_input.tif'
-        geo = 'admin2'
-        outf = f'outputs{sep}unittests'
+        mp = f"inputs{sep}test2_assetwealth_input.json"
+        fp = f"inputs{sep}test2_assetwealth_input.tif"
+        geo = "admin2"
+        outf = f"outputs{sep}unittests"
 
         # Process:
         df, dct = mixmasta.process(fp, mp, geo, outf)
-        #categories = df.select_dtypes(include=['category']).columns.tolist()
-        df['value'] = df['value'].astype('str')
+        # categories = df.select_dtypes(include=['category']).columns.tolist()
+        df["value"] = df["value"].astype("str")
 
         # Load expected output:
-        output_df = pd.read_csv(f'outputs{sep}test2_assetwealth_output.csv', index_col=False)
+        output_df = pd.read_csv(
+            f"outputs{sep}test2_assetwealth_output.csv", index_col=False
+        )
 
-        with open(f'outputs{sep}test2_assetwealth_dict.json') as f:
+        with open(f"outputs{sep}test2_assetwealth_dict.json") as f:
             output_dict = json.loads(f.read())
 
         # Sort both data frames and reindex for comparison,.
-        cols = ['timestamp','country','admin1','admin2','admin3','lat','lng','feature','value']
+        cols = [
+            "timestamp",
+            "country",
+            "admin1",
+            "admin2",
+            "admin3",
+            "lat",
+            "lng",
+            "feature",
+            "value",
+        ]
         df = df[cols]
         output_df = output_df[cols]
 
         # Optimize datatypes for output_df.
-        floats = output_df.select_dtypes(include=['float64']).columns.tolist()
-        output_df[floats] = output_df[floats].apply(pd.to_numeric, downcast='float')
+        floats = output_df.select_dtypes(include=["float64"]).columns.tolist()
+        output_df[floats] = output_df[floats].apply(pd.to_numeric, downcast="float")
 
-        ints = output_df.select_dtypes(include=['int64']).columns.tolist()
-        output_df[ints] = output_df[ints].apply(pd.to_numeric, downcast='integer')
+        ints = output_df.select_dtypes(include=["int64"]).columns.tolist()
+        output_df[ints] = output_df[ints].apply(pd.to_numeric, downcast="integer")
 
         # Standardize value and feature columns to str for comparison.
-        df['value'] = df['value'].astype('str')
-        df['feature'] = df['feature'].astype('str')
-        output_df['value'] = output_df['value'].astype('str')
-        output_df['feature'] = output_df['feature'].astype('str')
+        df["value"] = df["value"].astype("str")
+        df["feature"] = df["feature"].astype("str")
+        output_df["value"] = output_df["value"].astype("str")
+        output_df["feature"] = output_df["feature"].astype("str")
 
         # Sort and reindex.
         df.sort_values(by=cols, inplace=True)
         df.reset_index(drop=True, inplace=True)
         output_df.sort_values(by=cols, inplace=True)
-        output_df.reset_index(drop =True, inplace=True)
+        output_df.reset_index(drop=True, inplace=True)
 
         # Assertions
-        assert_frame_equal(df, output_df, check_categorical = False)
+        assert_frame_equal(df, output_df, check_categorical=False)
         assert_dict_equal(dct, output_dict)
 
     def test_003_process(self):
         """Test qualifies, lat/lng primary geo."""
 
         # Define mixmasta inputs:
-        mp = f'inputs{sep}test3_qualifies.json'
-        fp = f'inputs{sep}test3_qualifies.csv'
-        geo = 'admin2'
-        outf = f'outputs{sep}unittests'
+        mp = f"inputs{sep}test3_qualifies.json"
+        fp = f"inputs{sep}test3_qualifies.csv"
+        geo = "admin2"
+        outf = f"outputs{sep}unittests"
 
         # Process:
         df, dct = mixmasta.process(fp, mp, geo, outf)
 
         # Load expected output:
-        output_df = pd.read_csv(f'outputs{sep}test3_qualifies_output.csv', index_col=False)
+        output_df = pd.read_csv(
+            f"outputs{sep}test3_qualifies_output.csv", index_col=False
+        )
         output_df = mixmasta.optimize_df_types(output_df)
-        with open(f'outputs{sep}test3_qualifies_dict.json') as f:
+        with open(f"outputs{sep}test3_qualifies_dict.json") as f:
             output_dict = json.loads(f.read())
 
         # Sort both data frames and reindex for comparison,.
-        cols = ['timestamp','country','admin1','admin2','admin3','lat','lng','feature','value']
+        cols = [
+            "timestamp",
+            "country",
+            "admin1",
+            "admin2",
+            "admin3",
+            "lat",
+            "lng",
+            "feature",
+            "value",
+        ]
         df.sort_values(by=cols, inplace=True)
         output_df.sort_values(by=cols, inplace=True)
 
         df.reset_index(drop=True, inplace=True)
-        output_df.reset_index(drop =True, inplace=True)
+        output_df.reset_index(drop=True, inplace=True)
 
         # Make the datatypes the same for value/feature columns.
-        df['value'] = df['value'].astype('str')
-        df['feature'] = df['feature'].astype('str')
-        output_df['value'] = output_df['value'].astype('str')
-        output_df['feature'] = output_df['feature'].astype('str')
+        df["value"] = df["value"].astype("str")
+        df["feature"] = df["feature"].astype("str")
+        output_df["value"] = output_df["value"].astype("str")
+        output_df["feature"] = output_df["feature"].astype("str")
 
         # Assertions
-        assert_frame_equal(df, output_df, check_categorical = False)
+        assert_frame_equal(df, output_df, check_categorical=False)
         assert_dict_equal(dct, output_dict)
 
     def test_004_process(self):
         """Test .xlxs file, qualifies col with multi dtypes."""
 
         # Define mixmasta inputs:
-        mp = f'inputs{sep}test4_rainfall_error.json'
-        fp = f'inputs{sep}test4_rainfall_error.xlsx'
-        geo = 'admin2'
-        outf = f'outputs{sep}unittests'
+        mp = f"inputs{sep}test4_rainfall_error.json"
+        fp = f"inputs{sep}test4_rainfall_error.xlsx"
+        geo = "admin2"
+        outf = f"outputs{sep}unittests"
 
         # Process:
         df, dct = mixmasta.process(fp, mp, geo, outf)
 
         # Load expected output:
-        output_df = pd.read_csv(f'outputs{sep}test4_rainfall_error_output.csv', index_col=False)
+        output_df = pd.read_csv(
+            f"outputs{sep}test4_rainfall_error_output.csv", index_col=False
+        )
         output_df = mixmasta.optimize_df_types(output_df)
-        with open(f'outputs{sep}test4_rainfall_error_dict.json') as f:
+        with open(f"outputs{sep}test4_rainfall_error_dict.json") as f:
             output_dict = json.loads(f.read())
 
         # Sort both data frames and reindex for comparison,.
-        cols = ['timestamp','country','admin1','admin2','admin3','lat','lng','feature','value','MainCause']
+        cols = [
+            "timestamp",
+            "country",
+            "admin1",
+            "admin2",
+            "admin3",
+            "lat",
+            "lng",
+            "feature",
+            "value",
+            "MainCause",
+        ]
         df.sort_values(by=cols, inplace=True)
         output_df.sort_values(by=cols, inplace=True)
 
         df.reset_index(drop=True, inplace=True)
-        output_df.reset_index(drop =True, inplace=True)
+        output_df.reset_index(drop=True, inplace=True)
 
         # Make the datatypes the same for value/feature and qualifying columns.
-        df['value'] = df['value'].astype('str')
-        df['feature'] = df['feature'].astype('str')
-        df['MainCause'] = df['MainCause'].astype('str')
-        output_df['value'] = output_df['value'].astype('str')
-        output_df['feature'] = output_df['feature'].astype('str')
-        output_df['MainCause'] = output_df['MainCause'].astype('str')
+        df["value"] = df["value"].astype("str")
+        df["feature"] = df["feature"].astype("str")
+        df["MainCause"] = df["MainCause"].astype("str")
+        output_df["value"] = output_df["value"].astype("str")
+        output_df["feature"] = output_df["feature"].astype("str")
+        output_df["MainCause"] = output_df["MainCause"].astype("str")
 
         # Assertions
-        assert_frame_equal(df, output_df, check_categorical = False)
+        assert_frame_equal(df, output_df, check_categorical=False)
         assert_dict_equal(dct, output_dict)
 
     def test_005__command_line_interface(self):
@@ -215,26 +262,57 @@ class TestMixmaster(unittest.TestCase):
         """
 
         # Confirm CLI --help is available.
-        result = subprocess.run(['mixmasta', '--help'], capture_output=True, encoding='utf-8')
-        self.assertIn('Processor for generating CauseMos compliant datasets.', result.stdout)
-        self.assertIn('--help  Show this message and exit.', result.stdout)
+        result = subprocess.run(
+            ["mixmasta", "--help"], capture_output=True, encoding="utf-8"
+        )
+        self.assertIn(
+            "Processor for generating CauseMos compliant datasets.", result.stdout
+        )
+        self.assertIn("--help  Show this message and exit.", result.stdout)
         self.assertEqual(result.returncode, 0)
-        self.assertIn('causemosify-multi  Process multiple input files to generate a single',result.stdout)
-
+        self.assertIn(
+            "causemosify-multi  Process multiple input files to generate a single",
+            result.stdout,
+        )
 
         # Confirm CLI causemosify-multi --help is available.
-        result = subprocess.run(['mixmasta', "causemosify-multi", "--help"], capture_output=True, encoding='utf-8')
+        result = subprocess.run(
+            ["mixmasta", "causemosify-multi", "--help"],
+            capture_output=True,
+            encoding="utf-8",
+        )
         self.assertEqual(result.returncode, 0)
-        self.assertIn('Process multiple input files to generate a single CauseMos compliant',result.stdout)
+        self.assertIn(
+            "Process multiple input files to generate a single CauseMos compliant",
+            result.stdout,
+        )
 
         # Run causemosify-multi
-        inputs = "--inputs=[{\"input_file\": \"inputs" + f"{sep}test1_input.csv\",\"mapper\": \"inputs{sep}test1_input.json\"" + "},{\"input_file\": \""
-        inputs = inputs + f"inputs{sep}test3_qualifies.csv\",\"mapper\": \"inputs{sep}test3_qualifies.json\"" + "}]"
-        result = subprocess.run(['mixmasta', "causemosify-multi", inputs, "--geo=admin2", f"--output-file=outputs{sep}unittests"], capture_output=True, encoding='utf-8')
-        if (result.returncode != 0):
+        inputs = (
+            '--inputs=[{"input_file": "inputs'
+            + f'{sep}test1_input.csv","mapper": "inputs{sep}test1_input.json"'
+            + '},{"input_file": "'
+        )
+        inputs = (
+            inputs
+            + f'inputs{sep}test3_qualifies.csv","mapper": "inputs{sep}test3_qualifies.json"'
+            + "}]"
+        )
+        result = subprocess.run(
+            [
+                "mixmasta",
+                "causemosify-multi",
+                inputs,
+                "--geo=admin2",
+                f"--output-file=outputs{sep}unittests",
+            ],
+            capture_output=True,
+            encoding="utf-8",
+        )
+        if result.returncode != 0:
             print(result)
         self.assertEqual(result.returncode, 0)
-        
+
         ## Compare parquet files.
         df1 = pd.read_parquet(f"outputs{sep}unittests.1.parquet.gzip")
         df2 = pd.read_parquet(f"outputs{sep}unittests.2.parquet.gzip")
@@ -242,105 +320,149 @@ class TestMixmaster(unittest.TestCase):
         output_df = pd.read_parquet(f"outputs{sep}test5.parquet.gzip")
 
         # Sort both data frames and reindex for comparison,.
-        cols = ['timestamp','country','admin1','admin2','admin3','lat','lng','feature','value']
+        cols = [
+            "timestamp",
+            "country",
+            "admin1",
+            "admin2",
+            "admin3",
+            "lat",
+            "lng",
+            "feature",
+            "value",
+        ]
         df.sort_values(by=cols, inplace=True)
         output_df.sort_values(by=cols, inplace=True)
         df.reset_index(drop=True, inplace=True)
-        output_df.reset_index(drop =True, inplace=True)
+        output_df.reset_index(drop=True, inplace=True)
 
         logger.info(df.shape)
         logger.info(output_df.shape)
 
         # Assertion
-        assert_frame_equal(df, output_df, check_categorical = False)
+        assert_frame_equal(df, output_df, check_categorical=False)
 
         ## Compare str.parquet file.
         df = pd.read_parquet(f"outputs{sep}unittests_str.2.parquet.gzip")
         output_df = pd.read_parquet(f"outputs{sep}test5_str.parquet.gzip")
 
         # Sort both data frames and reindex for comparison,.
-        cols = ['timestamp','country','admin1','admin2','admin3','lat','lng','feature','value']
+        cols = [
+            "timestamp",
+            "country",
+            "admin1",
+            "admin2",
+            "admin3",
+            "lat",
+            "lng",
+            "feature",
+            "value",
+        ]
         df.sort_values(by=cols, inplace=True)
         output_df.sort_values(by=cols, inplace=True)
         df.reset_index(drop=True, inplace=True)
-        output_df.reset_index(drop =True, inplace=True)
+        output_df.reset_index(drop=True, inplace=True)
 
         # Assertion
-        assert_frame_equal(df, output_df, check_categorical = False)
+        assert_frame_equal(df, output_df, check_categorical=False)
 
     def test_006_process(self):
         """Test multi primary_geo, resolve_to_gadm"""
 
         # Define mixmasta inputs:
-        mp = f'inputs{sep}test6_hoa_conflict_input.json'
-        fp = f'inputs{sep}test6_hoa_conflict_input.csv'
-        geo = 'admin2'
-        outf = f'outputs{sep}unittests'
+        mp = f"inputs{sep}test6_hoa_conflict_input.json"
+        fp = f"inputs{sep}test6_hoa_conflict_input.csv"
+        geo = "admin2"
+        outf = f"outputs{sep}unittests"
 
         # Process:
         df, dct = mixmasta.process(fp, mp, geo, outf)
 
         # Load expected output:
-        output_df = pd.read_csv(f'outputs{sep}test6_hoa_conflict_output.csv', index_col=False)
+        output_df = pd.read_csv(
+            f"outputs{sep}test6_hoa_conflict_output.csv", index_col=False
+        )
         output_df = mixmasta.optimize_df_types(output_df)
-        with open(f'outputs{sep}test6_hoa_conflict_dict.json') as f:
+        with open(f"outputs{sep}test6_hoa_conflict_dict.json") as f:
             output_dict = json.loads(f.read())
 
         # Sort both data frames and reindex for comparison,.
-        cols = ['timestamp','country','admin1','admin2','admin3','lat','lng','feature','value']
+        cols = [
+            "timestamp",
+            "country",
+            "admin1",
+            "admin2",
+            "admin3",
+            "lat",
+            "lng",
+            "feature",
+            "value",
+        ]
         df.sort_values(by=cols, inplace=True)
         output_df.sort_values(by=cols, inplace=True)
 
         df.reset_index(drop=True, inplace=True)
-        output_df.reset_index(drop =True, inplace=True)
+        output_df.reset_index(drop=True, inplace=True)
 
         # Make the datatypes the same for value/feature and qualifying columns.
-        df['value'] = df['value'].astype('str')
-        df['feature'] = df['feature'].astype('str')
-        output_df['value'] = output_df['value'].astype('str')
-        output_df['feature'] = output_df['feature'].astype('str')
+        df["value"] = df["value"].astype("str")
+        df["feature"] = df["feature"].astype("str")
+        output_df["value"] = output_df["value"].astype("str")
+        output_df["feature"] = output_df["feature"].astype("str")
 
         # Assertions
-        assert_frame_equal(df, output_df, check_categorical = False)
+        assert_frame_equal(df, output_df, check_categorical=False)
         assert_dict_equal(dct, output_dict)
 
     def test_007_single_band_tif(self):
-        """ This tests single-band geotiff processing."""
+        """This tests single-band geotiff processing."""
 
         # Define mixmasta inputs:
-        mp = f'inputs{sep}test7_single_band_tif_input.json'
-        fp = f'inputs{sep}test7_single_band_tif_input.tif'
-        geo = 'admin2'
-        outf = f'outputs{sep}unittests'
+        mp = f"inputs{sep}test7_single_band_tif_input.json"
+        fp = f"inputs{sep}test7_single_band_tif_input.tif"
+        geo = "admin2"
+        outf = f"outputs{sep}unittests"
 
         # Process:
         df, dct = mixmasta.process(fp, mp, geo, outf)
-        #categories = df.select_dtypes(include=['category']).columns.tolist()
-        df['value'] = df['value'].astype('str')
+        # categories = df.select_dtypes(include=['category']).columns.tolist()
+        df["value"] = df["value"].astype("str")
 
         # Load expected output:
-        output_df = pd.read_csv(f'outputs{sep}test7_single_band_tif_output.csv', index_col=False)
+        output_df = pd.read_csv(
+            f"outputs{sep}test7_single_band_tif_output.csv", index_col=False
+        )
 
-        with open(f'outputs{sep}test7_single_band_tif_dict.json') as f:
+        with open(f"outputs{sep}test7_single_band_tif_dict.json") as f:
             output_dict = json.loads(f.read())
 
         # Sort both data frames and reindex for comparison,.
-        cols = ['timestamp','country','admin1','admin2','admin3','lat','lng','feature','value']
+        cols = [
+            "timestamp",
+            "country",
+            "admin1",
+            "admin2",
+            "admin3",
+            "lat",
+            "lng",
+            "feature",
+            "value",
+        ]
         df = df[cols]
         output_df = output_df[cols]
 
         # Optimize datatypes for output_df.
-        floats = output_df.select_dtypes(include=['float64']).columns.tolist()
-        output_df[floats] = output_df[floats].apply(pd.to_numeric, downcast='float')
+        floats = output_df.select_dtypes(include=["float64"]).columns.tolist()
+        output_df[floats] = output_df[floats].apply(pd.to_numeric, downcast="float")
 
-        ints = output_df.select_dtypes(include=['int64']).columns.tolist()
-        output_df[ints] = output_df[ints].apply(pd.to_numeric, downcast='integer')
+        ints = output_df.select_dtypes(include=["int64"]).columns.tolist()
+        output_df[ints] = output_df[ints].apply(pd.to_numeric, downcast="integer")
 
         # Standardize value and feature columns to str for comparison.
-        df['value'] = df['value'].astype('str')
-        df['feature'] = df['feature'].astype('str')
-        output_df['value'] = output_df['value'].astype('str')
-        output_df['feature'] = output_df['feature'].astype('str')
+        df["value"] = df["value"].astype("str")
+        df["feature"] = df["feature"].astype("str")
+        output_df["value"] = output_df["value"].astype("str")
+        output_df["feature"] = output_df["feature"].astype("str")
 
         # Sort and reindex.
         df.sort_values(by=cols, inplace=True)
@@ -350,22 +472,36 @@ class TestMixmaster(unittest.TestCase):
         output_df.reset_index(drop=True, inplace=True)
 
         # Assertions
-        assert_frame_equal(df, output_df, check_categorical = False)
+        assert_frame_equal(df, output_df, check_categorical=False)
         assert_dict_equal(dct, output_dict)
 
     def test_008_aliases(self):
-        """ This tests feature name aliases."""
+        """This tests feature name aliases."""
 
         # Define mixmasta inputs:
-        mp = f'inputs{sep}test8_aliases_input.json'
-        fp = f'inputs{sep}test8_aliases_input.csv'
-        geo = 'admin2'
-        outf = f'outputs{sep}unittests'
+        mp = f"inputs{sep}test8_aliases_input.json"
+        fp = f"inputs{sep}test8_aliases_input.csv"
+        geo = "admin2"
+        outf = f"outputs{sep}unittests"
 
-        inputs = "--inputs=[{\"input_file\": \"inputs" + f"{sep}test8_aliases_input.csv\",\"mapper\": \"inputs{sep}test8_aliases_input.json\"" + "}]"
-        result = subprocess.run(['mixmasta', "causemosify-multi", inputs, "--geo=admin2", f"--output-file=outputs{sep}unittests"], capture_output=True, encoding='utf-8')
+        inputs = (
+            '--inputs=[{"input_file": "inputs'
+            + f'{sep}test8_aliases_input.csv","mapper": "inputs{sep}test8_aliases_input.json"'
+            + "}]"
+        )
+        result = subprocess.run(
+            [
+                "mixmasta",
+                "causemosify-multi",
+                inputs,
+                "--geo=admin2",
+                f"--output-file=outputs{sep}unittests",
+            ],
+            capture_output=True,
+            encoding="utf-8",
+        )
 
-        if (result.returncode != 0):
+        if result.returncode != 0:
             print(result)
         self.assertEqual(result.returncode, 0)
 
@@ -379,10 +515,10 @@ class TestMixmaster(unittest.TestCase):
         output_df = output_df_1.append(output_df_2)
 
         # Assertions
-        assert_frame_equal(df, output_df, check_categorical = False)
+        assert_frame_equal(df, output_df, check_categorical=False)
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     unittest.main()
 
 """
